requirements.txt
flask==3.0.3
gunicorn==21.2.0
stripe==9.9.0
requests==2.31.0
beautifulsoup4==4.12.3
google-generativeai==0.5.4
APScheduler==3.10.4
colorlog==6.8.2
python-dotenv==1.0.1
lxml==5.2.1
